cheri: Ensure that memory for smbios structures is mapped correctly.
The smbios structures have a string table appended after the structure contents.
The memory is initialised using the map_sysmem() call, however the length passed
to this function does not take into account the size of the string table.
Since it is not possible to precompute this size, use an arbitrary maximum value
that is checked in smbios_add_string for consistency.